% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createShinyBurstsView}
\alias{createShinyBurstsView}
\title{Create a materialized view of all bursts for the shiny app}
\usage{
createShinyBurstsView(conn, schema)
}
\arguments{
\item{conn}{DBI::DBIConnection}

\item{schema}{String. Schema name.}
}
\value{
nothing
}
\description{
It is expected that *all* pgtrajes are projected in the schema in order to
run. Bursts are projected to EPSG:4326 thus there is no need for coordinate transformation
for leaflet.
}
\author{
Balázs Dukai \email{balazs.dukai@gmail.com}
}
\keyword{internal}
