% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createShinyStepsView}
\alias{createShinyStepsView}
\title{Create a materialized view of the steps of a pgtraj for the shiny app}
\usage{
createShinyStepsView(conn, schema, pgtraj)
}
\arguments{
\item{conn}{DBI::DBIConnection}

\item{schema}{String. Schema name.}

\item{pgtraj}{String. Pgtraj name.}
}
\value{
nothing
}
\description{
Steps are projected to EPSG:4326 thus there is no need for coordinate transformation
for leaflet.
}
\author{
Balázs Dukai \email{balazs.dukai@gmail.com}
}
\keyword{internal}
