% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createShinyViews.R
\name{createShinyViews}
\alias{createShinyViews}
\title{Create database views that are suitable for explorePgtraj()}
\usage{
createShinyViews(conn, schema, pgtraj, force = FALSE)
}
\arguments{
\item{conn}{DBI::DBIConnection}

\item{schema}{String. Schema name.}

\item{pgtraj}{String. Pgtraj name.}

\item{force}{Boolean. Drop and recreate the views if they already exist.}
}
\value{
nothing
}
\description{
It is expected that *all* pgtrajes are projected in the schema in order to
run.
}
\examples{
\dontrun{
createShinyViews(conn, schema="ibex_traj", pgtraj="ibex", force=TRUE)
}
}
