% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shiny.R
\name{findGeoType}
\alias{findGeoType}
\title{Figures out whether the provided database relation contains vector or raster data.}
\usage{
findGeoType(conn, layers)
}
\arguments{
\item{conn}{DBI::DBIConnection}

\item{layers}{List. List of character vectors for each layer to include as a 
base layer.}
}
\value{
List of lists of database relations as \code{list(vect = list(), rast = list())}.
}
\description{
Not implemented for rasters.
}
\examples{
\dontrun{
layers <- list(c("example_data", "county_subdiv"), c("public", "florida_dem"))
geo_type <- findGeoType(conn, layers)
geo_type$vect[[1]]
}
}
\author{
Balázs Dukai \email{balazs.dukai@gmail.com}
}
\keyword{internal}
