% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shiny.R
\name{getAnimalsDf}
\alias{getAnimalsDf}
\title{Get distinct animal names in step_geometry view}
\usage{
getAnimalsDf(conn, schema, view)
}
\arguments{
\item{conn}{DBI::DBIConnection}

\item{schema}{String. Schema name.}

\item{view}{String. View name.}
}
\value{
data frame with column 'animal_name'
}
\description{
Get distinct animal names in step_geometry view
}
\author{
Balázs Dukai \email{balazs.dukai@gmail.com}
}
\keyword{internal}
