% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getBurstGeom}
\alias{getBurstGeom}
\title{Get geometry of bursts as linestring}
\usage{
getBurstGeom(conn, schema, view, burst_name)
}
\arguments{
\item{conn}{DBI::DBIConnection}

\item{schema}{String. Schema name.}

\item{view}{String. View name.}

\item{burst_name}{String. Accepts a character vector of variable length}
}
\value{
a single LINESTRING per burst
}
\description{
Get geometry of bursts as linestring
}
\author{
Balázs Dukai \email{balazs.dukai@gmail.com}
}
\keyword{internal}
