% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shiny.R
\name{getBurstsDF}
\alias{getBurstsDF}
\title{Get distinct burst names in a step_geometry view}
\usage{
getBurstsDF(conn, schema, view)
}
\arguments{
\item{conn}{DBI::DBIConnection}

\item{schema}{String. Schema name.}

\item{view}{String. View name.}
}
\value{
data frame with column 'burst_name'
}
\description{
Get distinct burst names in a step_geometry view
}
\author{
Balázs Dukai \email{balazs.dukai@gmail.com}
}
\keyword{internal}
