% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shiny.R
\name{getFullTraj}
\alias{getFullTraj}
\title{Get the complete trajectory of an animal as a single linestring}
\usage{
getFullTraj(conn, schema, view)
}
\arguments{
\item{conn}{DBI::DBIConnection}

\item{schema}{String. Schema name.}

\item{view}{String. View name.}
}
\value{
a single LINESTRING per animal
}
\description{
Get the complete trajectory of an animal as a single linestring
}
\author{
Balázs Dukai \email{balazs.dukai@gmail.com}
}
\keyword{internal}
