% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getInfolocsColumns}
\alias{getInfolocsColumns}
\title{Get all columns in the infolocs table but the step_id}
\usage{
getInfolocsColumns(conn, schema, pgtraj, df = FALSE)
}
\arguments{
\item{conn}{DBI::DBIConnection}

\item{schema}{String. Schema name.}

\item{pgtraj}{String. Pgtraj name.}

\item{df}{Boolean. Return a data frame or a string?}
}
\value{
character vector or NULL if there are no infolocs
}
\description{
Gets all the columns names in the infoloc table of the pgtraj and parses
them for inserting into an SQL query, e.g.: "col1 ,col2, col2 ,"
}
\author{
Balázs Dukai \email{balazs.dukai@gmail.com}
}
\keyword{internal}
