% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shiny.R
\name{getLayers}
\alias{getLayers}
\title{Get base layers from database}
\usage{
getLayers(conn, layers)
}
\arguments{
\item{conn}{DBI::DBIConnection}

\item{layers}{List. List of character vectors for each layer to include as a
base layer.}
}
\value{
list of simple features as \code{list(name=sf object, name2=sf object)}
}
\description{
Not implemented for rasters. Transforms coordinates to EPSG:4326.
}
\examples{
\dontrun{
layers <- list(c("schema1", "tableA"), c("schema2", "tableB"))
}
}
\author{
Balázs Dukai \email{balazs.dukai@gmail.com}
}
\keyword{internal}
