% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPgtrajWithInfo}
\alias{getPgtrajWithInfo}
\title{Get pgtraj with infolocs (and R class definitions, if they exist)}
\usage{
getPgtrajWithInfo(conn, pgtraj, schema)
}
\description{
Get pgtraj with infolocs (and R class definitions, if they exist)
}
\keyword{internal}

