% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getStepWindow}
\alias{getStepWindow}
\title{Get steps within a temporal window}
\usage{
getStepWindow(conn, schema, view, time, interval, step_mode, info_cols,
  tstamp_start, tstamp_last)
}
\arguments{
\item{conn}{DBI::DBIConnection}

\item{schema}{String. Schema name.}

\item{view}{String. View name.}

\item{time}{String of the start time of the time window. Including time zone.}

\item{interval}{lubridate::lubridate::period object of the time window}

\item{step_mode}{Boolean. Detailed step info (TRUE) or aggregate}

\item{info_cols}{Character vector of the infolocs columns of the pgtraj.}

\item{tstamp_start}{POSIXct with timestamp. First time stamp in view.}

\item{tstamp_last}{POSIXct with timestamp. Last time stamp in view.}
}
\value{
A simple feature object of the steps. NULL when out of range.
}
\description{
Get steps within a temporal window
}
\author{
Balázs Dukai \email{balazs.dukai@gmail.com}
}
\keyword{internal}
