% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shiny.R
\name{getTrajDefaults}
\alias{getTrajDefaults}
\title{Get default time parameters for steps}
\usage{
getTrajDefaults(conn, schema, view, pgtraj)
}
\arguments{
\item{conn}{DBI::DBIConnection}

\item{schema}{String. Schema name.}

\item{view}{String. View name.}

\item{pgtraj}{String. Pgtraj name}
}
\value{
data frame with columns: tstamp_start (epoch), tstamp_last (epoch), increment, tzone
}
\description{
Get default time parameters for steps
}
\author{
Balázs Dukai \email{balazs.dukai@gmail.com}
}
\keyword{internal}
