% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_blank}
\alias{is_blank}
\title{Test if an argument is either NA, NULL, NaN or empty string.}
\usage{
is_blank(x, false_triggers = FALSE)
}
\arguments{
\item{x}{Any object.}
}
\value{
boolean
}
\description{
Test if an argument is either NA, NULL, NaN or empty string.
}
\author{
Balázs Dukai \email{balazs.dukai@gmail.com}
}
\keyword{internal}
