% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shiny.R
\name{isVector}
\alias{isVector}
\title{Does a table contain vector data?}
\usage{
isVector(conn, layer)
}
\arguments{
\item{conn}{DBI::DBIConnection}

\item{layer}{String. As c(schema, table)}
}
\value{
Boolean
}
\description{
Does a table contain vector data?
}
\author{
Balázs Dukai \email{balazs.dukai@gmail.com}
}
\keyword{internal}
