% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_opt.R
\name{ld_opt}
\alias{dl_opt}
\alias{ld_opt}
\title{Quick Conversion of Objects of Class ltraj from and to Dataframes}
\usage{
ld_opt(ltraj)

dl_opt(x)
}
\arguments{
\item{ltraj}{An object of class \code{ltraj}.}
}
\description{
Faster versions of \code{\link[adehabitatLT]{ld}} and
\code{\link[adehabitatLT]{dl}}.
}
\author{
Modified by Mathieu Basille \email{basille@ufl.edu},
    Bal<U+FFFD><U+FFFD>zs Dukai \email{balazs.dukai@gmail.com}
}
\seealso{
See \code{\link[adehabitatLT]{ld}} for further details on
    the function and all available arguments.
}
\keyword{internal}

