% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltraj2pgtraj.R
\name{ltraj2pgtraj}
\alias{ltraj2pgtraj}
\title{Export an ltraj object from R into a traj database schema.}
\usage{
ltraj2pgtraj(conn, ltraj, schema = "traj", pgtraj = NULL, epsg = NULL,
  comment = NULL)
}
\arguments{
\item{conn}{Connection object created with RPostgreSQL}

\item{ltraj}{An ltraj object.}

\item{schema}{String. Name of the schema that stores or will store the pgtraj data model.}

\item{pgtraj}{String. Name of the new pgtraj. Defaults to the name of the 
variable that stores the ltraj.}

\item{epsg}{Numeric. The EPSG code of the Coordinate Reference System of the 
relocation coordinates in the ltraj. Defaults to 0.}

\item{comment}{String. A comment that will be stored with the pgtraj in the database.}
}
\value{
TRUE on success
}
\description{
\code{ltraj2pgtraj} creates a new traj schema or uses an existing one and 
exports an ltraj to the database. Uses \code{as_pgtraj} to insert the values 
into the traj schema.
}
\examples{
\dontrun{ltraj2pgtraj(conn, ibex, "traj_t2")}

}
\author{
Balázs Dukai \email{balazs.dukai@gmail.com}
}
\seealso{
\code{\link{as_pgtraj}}
}

