% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgtraj2ltraj.R
\name{pgtraj2ltraj}
\alias{pgtraj2ltraj}
\title{Import a pgtraj into an ltraj.}
\usage{
pgtraj2ltraj(conn, schema = "traj", pgtraj)
}
\arguments{
\item{conn}{Connection object created with RPostgreSQL}

\item{schema}{String. Name of the schema that stores or will store the pgtraj data model.}

\item{pgtraj}{String. Name of the pgtraj.}
}
\value{
an ltraj object
}
\description{
\code{pgtraj2ltraj} imports a single pgtraj from a database into an ltraj object.
}
\examples{
\dontrun{pgtraj2ltraj(conn, "traj_t2", "ibex")}

}
\author{
Bal<U+FFFD><U+FFFD>zs Dukai \email{balazs.dukai@gmail.com}
}

