% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgTrajDelete.R
\name{pgTrajDelete}
\alias{pgTrajDelete}
\title{Delete one or more pgtrajs from a traj schema.}
\usage{
pgTrajDelete(conn, schema = "traj", pgtraj)
}
\arguments{
\item{conn}{Connection object created with RPostgreSQL}

\item{schema}{String. Name of the schema that stores the traj data model.}

\item{pgtraj}{String. A vector containing the names of the pgtrajs.}
}
\value{
nothing
}
\description{
\code{pgTrajDelete} deletes one or more pgtrajs from a traj schema.
}
\examples{
\dontrun{pgTrajDelete(conn, "traj", "ibex")}

\dontrun{pgTrajDelete(conn, schema="traj", pgtraj=c("ibex", "puechcirc")}

}
\author{
Bal<U+FFFD><U+FFFD>zs Dukai \email{balazs.dukai@gmail.com}
}

