% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgTrajDrop.R
\name{pgTrajDrop}
\alias{pgTrajDrop}
\title{Delete one or more pgtrajs from a traj schema.}
\usage{
pgTrajDrop(conn, pgtraj, schema = "traj")
}
\arguments{
\item{conn}{Connection object created with RPostgreSQL}

\item{pgtraj}{String. A vector containing the names of the pgtrajs.}

\item{schema}{String. Name of the schema that stores the traj data model.}
}
\value{
TRUE on success
}
\description{
\code{pgTrajDrop} deletes one or more pgtrajs from a traj schema.
}
\examples{
\dontrun{
  # drop "ibex" pgtraj in schema "traj"
  pgTrajDrop(conn, "ibex")
  # drop "ibex" pgtraj and "puechcirc" pgtraj in schema "traj"
  pgTrajDrop(conn, pgtraj=c("ibex", "puechcirc")
}
}
\author{
Balázs Dukai \email{balazs.dukai@gmail.com}
}

