% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgTrajR2TempT.R
\name{pgTrajR2TempT}
\alias{pgTrajR2TempT}
\title{Insert an ltraj data frame into the 'qqbqahfsbrpq_temp' table.
Input is an ltraj converted into a data frame with ld_opt().
Ltraj row names are preserved. No transaction control.}
\usage{
pgTrajR2TempT(conn, schema, dframe, pgtraj, srid = 0)
}
\arguments{
\item{conn}{Connection object created with RPostgreSQL}

\item{schema}{String. Name of the schema that stores or will store the pgtraj data model.}

\item{dframe}{Data frame created from an ltraj with the function ld_opt().}

\item{pgtraj}{String. Name of the new pgtraj. Defaults to the name of the 
variable that stores the ltraj.}

\item{srid}{Numeric. The PostGIS SRID of the Coordinate Reference System of the 
relocation coordinates in the ltraj. Defaults to 0.}
}
\value{
TRUE on success, otherwise warning/error
}
\description{
Insert an ltraj data frame into the 'qqbqahfsbrpq_temp' table.
Input is an ltraj converted into a data frame with ld_opt().
Ltraj row names are preserved. No transaction control.
}
\note{
ST_PointFromText() vs. ST_MakePoint() http://gis.stackexchange.com/a/122263/56083
}

