% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgTrajViews.R
\name{pgTrajViewParams}
\alias{pgTrajViewParams}
\title{Computes the trajectory parameters (as in ltraj) for a pgtraj and creates a 
view for the pgtraj. The views are always named as '<pgtraj>_params'.}
\usage{
pgTrajViewParams(conn, schema, pgtraj, epsg)
}
\arguments{
\item{conn}{Connection object created with RPostgreSQL}

\item{schema}{String. Name of the schema that stores or will store the pgtraj data model.}

\item{pgtraj}{String. Name of the pgtraj.}

\item{epsg}{Numeric. EPSG code of the relocation geometry.}
}
\value{
TRUE on success, otherwise warning/error
}
\description{
Computes the trajectory parameters (as in ltraj) for a pgtraj and creates a 
view for the pgtraj. The views are always named as '<pgtraj>_params'.
}
\author{
Bal<U+FFFD><U+FFFD>zs Dukai
}

