% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgTrajViews.R
\name{pgTrajViewStepGeom}
\alias{pgTrajViewStepGeom}
\title{Creates a view of the step geometries for visualization.}
\usage{
pgTrajViewStepGeom(conn, schema, pgtraj)
}
\arguments{
\item{conn}{Connection object created with RPostgreSQL}

\item{schema}{String. Name of the schema that stores or will store the pgtraj data model.}
}
\value{
TRUE on success, otherwise warning/error
}
\description{
Creates a view of the step geometries for visualization.
}
\author{
Bal<U+FFFD><U+FFFD>zs Dukai
}

