% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpostgisLT-package.R
\docType{package}
\name{rpostgisLT}
\alias{rpostgisLT}
\alias{rpostgisLT-package}
\title{Integration of ltraj (adehabitatLT) and pgtraj (PostGIS).}
\description{
rpostgisLT
}
\details{
The `rpostgisLT` package develops the integration of R and PostGIS
for managing movement trajectories. The focus is on streamlining
the workflow for biologists to store and process animal
trajectories in PostGIS and analyze them in R, thus utilizing the
strengths of both software. The package relies on `ltraj` objects
from the R package `adehabitatLT`, and provides the analogous
`pgtraj` data structure in PostGIS, with all functions to create
and manage `pgtraj` data, and convert from and to both format
(`pgtraj` in PostGIS, `ltraj` in R). For a list of documented
functions, use \code{library(help = "rpostgisLT")}
}
\author{
Balázs Dukai \email{balazs.dukai@gmail.com}
}

