% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trajSummaryViews}
\alias{trajSummaryViews}
\title{Create views that summarize all pgtrajes (and pgtraj bursts) in a schema.}
\usage{
trajSummaryViews(conn, schema)
}
\arguments{
\item{conn}{A PostgreSQL connection object}

\item{schema}{String, name of the schema}
}
\description{
Create views that summarize all pgtrajes (and pgtraj bursts) in a schema.
}
\author{
David Bucklin \email{dbucklin@ufl.edu}
}
\keyword{internal}

