% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateNumericTimeInput}
\alias{updateNumericTimeInput}
\title{Convert the value of input$interval/increment to the unit selected in input$*_unit}
\usage{
updateNumericTimeInput(session, inputUnit, inputId, reactiveTime)
}
\arguments{
\item{session}{Shiny session}

\item{inputUnit}{String. One of years, months, days, hours, minutes seconds}

\item{inputId}{String. Id of the input slot.}

\item{reactiveTime}{A lubridate::lubridate::period object stored in a Reactive Value}
}
\value{
nothing
}
\description{
Convert the value of input$interval/increment to the unit selected in input$*_unit
}
\author{
Balázs Dukai \email{balazs.dukai@gmail.com}
}
\keyword{internal}
