% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeInfoFromDB}
\alias{writeInfoFromDB}
\title{Write infolocs table to database from database in pgtraj2ltraj()}
\usage{
writeInfoFromDB(conn, pgtraj, schema, info_cols, info_table, info_rids)
}
\arguments{
\item{conn}{A PostgreSQL connection object.}

\item{pgtraj}{String, name of the pgtraj}

\item{schema}{String, name of the schema holding the pgtraj}

\item{info_cols}{String. Optional character vector of column names of 
additional information on relocations (replicating "infolocs" from the
\code{adehabitatLT} object \code{ltraj}).}

\item{info_table}{Character vector of \code{c("schema","table")} holding the 
\code{info_cols}.}

\item{info_rids}{String. Column name of unique integer ID in \code{info_table} to join with
\code{rids} of the original relocations.}
}
\value{
TRUE on successful infolocs writing
}
\description{
Write infolocs table to database from database in pgtraj2ltraj()
}
\author{
David Bucklin \email{dbucklin@ufl.edu}
}
\keyword{internal}

