% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeInfoFromLtraj}
\alias{writeInfoFromLtraj}
\title{Write infolocs table to database from ltraj in ltraj2pgtraj()}
\usage{
writeInfoFromLtraj(conn, ltraj, pgtraj, schema)
}
\arguments{
\item{conn}{A PostgreSQL connection object.}

\item{ltraj}{An adehabitatLT ltraj object}

\item{pgtraj}{String, name of the pgtraj being created}

\item{schema}{String, name of the schema holding the pgtraj}
}
\value{
TRUE on successful infolocs writing
}
\description{
Write infolocs table to database from ltraj in ltraj2pgtraj()
}
\author{
David Bucklin \email{dbucklin@ufl.edu}
}
\keyword{internal}
