% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_pgtraj.R
\name{as_pgtraj}
\alias{as_pgtraj}
\title{Imports location data from a database table into a 'traj' schema.}
\usage{
as_pgtraj(conn, relocations_table, schema = "traj", pgtrajs = "pgtraj",
  animals = "animal", bursts = NULL, relocations, timestamps = NULL,
  rids = "rid", srid = NULL, note = NULL, info_cols = NULL,
  info_table = NULL, info_rids = NULL)
}
\arguments{
\item{conn}{Connection object created with RPostgreSQL}

\item{relocations_table}{String. Name of the schema and table that stores the relocations, e.g. c("schema","relocations")}

\item{schema}{String. Name of the schema that stores or will store the pgtraj data model.}

\item{pgtrajs}{String. Name of the pgtraj or name of the field that stores the pgtraj names.}

\item{animals}{String. Name of the animal or name of the field that stores the animal names.}

\item{bursts}{String. Name of the burst or name of the field that stores the burst names.}

\item{relocations}{String. Name of the field that contains the relocations 
in relocations_table. Relocations can be provided either as X,Y coordinates
or PostGIS geometry. In both cases all relocations in the 'relocations_table'
have to have the same projection.}

\item{timestamps}{String. Name of the field in relocations_table that contains the timestamps.
If NULL, Type I trajectory is assumed.}

\item{rids}{String. Name of the field in relocations_table that contains the numeric IDs of relocations.}

\item{srid}{Integer. Optional SRID (spatial reference ID) of (x,y) coordinates provided for relocations.
Ignored if relocations is a geometry type.}

\item{note}{String. Comment on the pgtraj. The comment is only used in
the database and not transferred into an ltraj.}

\item{info_cols}{String. Optional character vector of column names of 
additional information on relocations (replicating "infolocs" from the
\code{adehabitatLT} object \code{ltraj}).}

\item{info_table}{Character vector of \code{c("schema","table")} holding the 
\code{info_cols}. If \code{info_cols} are in \code{relocations_table}, leave NULL.}

\item{info_rids}{String. Column name of unique integer ID in \code{info_table} to join with
\code{rids}. If \code{info_cols} are in \code{relocations_table}, leave NULL.}
}
\value{
TRUE on success
}
\description{
\code{as_pgtraj} populates a \code{traj} schema from the data provided
in \code{relocations_table}. If the provided schema doesn't exist, it will be 
created. On successful data input, \code{as_pgtraj} creates three database views for
each pgtraj. These views are named <pgtraj_name>_parameters, 
<pgtraj_name>_step_geometry and <pgtraj_name>_summary and described more in
detail in the package vignette.

The time zone of the pgtraj is set to the local time zone of the user.
}
\details{
Opening and closing connections have to be done manually by the user. 
However, the function checks if the provided connection is still valid. 
Not tested with capital letters for PostgreSQL field names.
}
\examples{
\dontrun{
as_pgtraj(conn, 
        relocations_table = c("example_data","relocations_plus"),
        schema = "traj_t4",
        pgtrajs = "id",
        animals = "animal",
        bursts = "burst",
        relocations = "geom",
        timestamps = "time",
        rids = "gid")
}

\dontrun{
as_pgtraj(conn, 
        relocations_table = c("example_data","relocations_plus"),
        schema = "traj_t4",
        pgtrajs = "id",
        animals = "animal",
        bursts = "burst",
        relocations = c("x","y"),
        timestamps = "time",
        rids = "gid")
}

}
\author{
Balázs Dukai \email{balazs.dukai@gmail.com}
}
\references{
\url{https://cran.r-project.org/web/packages/adehabitatLT/vignettes/adehabitatLT.pdf}
}
\seealso{
Section on traj data model in the package vignette.
}

