% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPgtrajWithInfo}
\alias{getPgtrajWithInfo}
\title{Get pgtraj with infolocs as a data frame (used in pgtraj2ltraj)}
\usage{
getPgtrajWithInfo(conn, pgtraj, schema)
}
\arguments{
\item{conn}{A PostgreSQL connection object.}

\item{pgtraj}{String, name of the pgtraj}

\item{schema}{String, name of the schema holding the pgtraj}
}
\description{
Get pgtraj with infolocs as a data frame (used in pgtraj2ltraj)
}
\author{
David Bucklin \email{dbucklin@ufl.edu}
}
\keyword{internal}

