% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltraj2pgtraj.R
\name{ltraj2pgtraj}
\alias{ltraj2pgtraj}
\title{Export an ltraj object from R into a traj database schema.}
\usage{
ltraj2pgtraj(conn, ltraj, schema = "traj", pgtraj = NULL, comment = NULL,
  create = FALSE, new.srid = NULL)
}
\arguments{
\item{conn}{Connection object created with RPostgreSQL}

\item{ltraj}{An ltraj object.}

\item{schema}{String. Name of the schema that stores or will store the pgtraj data model.}

\item{pgtraj}{String. Name of the new pgtraj. Defaults to the name of the 
variable that stores the ltraj.}

\item{comment}{String. A comment that will be stored with the pgtraj in the database.}

\item{create}{Logical. If no matching SRID is found, should a new SRID be created? 
User must have write access on spatial_ref_sys table.}

\item{new.srid}{Integer. Optional SRID to give to a newly created SRID. If left NULL (default),
the next open value of `srid` in `spatial_ref_sys` between 880000 and 890000 will be used.}
}
\value{
TRUE on success
}
\description{
\code{ltraj2pgtraj} creates a new traj schema or uses an existing one and 
exports an ltraj to the database. The time zone and projection information
stored in the ltraj is transferred to the database. Uses \code{as_pgtraj} 
to insert the values into the traj schema.
}
\examples{
\dontrun{ltraj2pgtraj(conn, ibex, "traj_t2")}

}
\author{
Bal<U+FFFD><U+FFFD>zs Dukai \email{balazs.dukai@gmail.com}
}
\seealso{
\code{\link{as_pgtraj}}
}

