% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgTrajDrop.R
\name{pgTrajDrop}
\alias{pgTrajDrop}
\title{Delete one or more pgtrajs from a traj schema.}
\usage{
pgTrajDrop(conn, schema = "traj", pgtraj)
}
\arguments{
\item{conn}{Connection object created with RPostgreSQL}

\item{schema}{String. Name of the schema that stores the traj data model.}

\item{pgtraj}{String. A vector containing the names of the pgtrajs.}
}
\value{
TRUE on success
}
\description{
\code{pgTrajDrop} deletes one or more pgtrajs from a traj schema.
}
\examples{
\dontrun{pgTrajDrop(conn, "traj", "ibex")}

\dontrun{pgTrajDrop(conn, schema="traj", pgtraj=c("ibex", "puechcirc")}

}
\author{
Balázs Dukai \email{balazs.dukai@gmail.com}
}

