% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgtrajDrop.R
\name{pgtrajDrop}
\alias{pgtrajDrop}
\title{Delete a pgtraj from a traj schema.}
\usage{
pgtrajDrop(conn, pgtraj, schema = "traj", full_clean = TRUE)
}
\arguments{
\item{conn}{Connection object created with RPostgreSQL}

\item{pgtraj}{String. Name of the pgtraj.}

\item{schema}{String. Name of the schema that stores the traj data model.}

\item{full_clean}{String. Whether to delete all unused rows in all tables
associated with the pgtraj (and any other previously deleted pgtrajs). Recommended, 
but may take a long time to run in schemas with many large pgtraj's.}
}
\value{
TRUE on success
}
\description{
\code{pgtrajDrop} deletes a pgtraj from a traj schema.
}
\examples{
\dontrun{
  # drop "ibex" pgtraj in schema "traj"
  pgtrajDrop(conn, "ibex")
}
}
\author{
Balázs Dukai \email{balazs.dukai@gmail.com}
}

