% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgTrajSchema.R
\name{pgTrajSchema}
\alias{pgTrajSchema}
\title{Check 'traj' schema.}
\usage{
pgTrajSchema(conn, name = "traj")
}
\arguments{
\item{conn}{Connection object created with RPostgreSQL.}

\item{schema}{Chaaracter string. Name of the schema that stores or
will store the pgtraj data model.}
}
\value{
TRUE on success
}
\description{
Checks if the provided schema exists in the database, and creates
if it doesn't.
}
\details{
Creates a 'traj' schema in the database by calling a SQL script 
from (./inst/sql/). The schema name defaults to 'traj'.
}
\examples{
\dontrun{pgTrajSchema(conn, "traj_1")}
}
\author{
Balázs Dukai \email{balazs.dukai@gmail.com}
}

