% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pgTrajViewStepGeom}
\alias{pgTrajViewStepGeom}
\title{Creates a view of the step geometries for visualization.}
\usage{
pgTrajViewStepGeom(conn, schema, pgtraj)
}
\arguments{
\item{conn}{Connection object created with RPostgreSQL}

\item{schema}{String. Name of the schema that stores or will store the 
pgtraj data model.}

\item{pgtraj}{String. Name of the pgtraj.}
}
\value{
TRUE on success, otherwise warning/error
}
\description{
Creates a view of the step geometries for visualization.
}
\author{
Balázs Dukai
}
\keyword{internal}

