% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setTimeInput}
\alias{setTimeInput}
\title{Set a lubridate::lubridate::period value from input$interval/increment}
\usage{
setTimeInput(inputUnit, inputTime, reactiveTime)
}
\arguments{
\item{inputUnit}{String.}

\item{inputTime}{lubridate::lubridate::period}

\item{reactiveTime}{Reactive value to set}
}
\value{
nothing
}
\description{
Set a lubridate::lubridate::period value from input$interval/increment
}
\author{
Balázs Dukai \email{balazs.dukai@gmail.com}
}
\keyword{internal}
